zscanner: revert changes committed by mistake

+++ src/zscanner/scanner.c
@@ -5204,7 +5204,7 @@ _resume:
 			else if ( _widec > *_mid )
 				_lower = _mid + 1;
 			else {
-				_trans += (_mid - _keys);
+				_trans += (unsigned int)(_mid - _keys);
 				goto _match;
 			}
 		}
@@ -5227,7 +5227,7 @@ _resume:
 			else if ( _widec > _mid[1] )
 				_lower = _mid + 2;
 			else {
-				_trans += ((_mid - _keys)>>1);
+				_trans += (unsigned int)((_mid - _keys)>>1);
 				goto _match;
 			}
 		}
